innitial commit

+++ src/main/java/net/javaguides/banking_app/dto/AccountDto.java
@@ -2,5 +2,5 @@ package net.javaguides.banking_app.dto;
 
 public record AccountDto(Long id,
                          String accountHolderName,
-                         Long balance) {
+                         Double balance) {
 }
